+++ demisto_sdk/commands/postman_codegen/tests/postman_codegen_test.py
@@ -250,15 +250,15 @@ class TestPostmanCodeGen:
     def test_args_lowercase(self, tmp_path):
         """
         Given
-        - postman collection
-        - with request Test Report which has variables with upper case
+        - Postman collection.
+        - Test report request which has variables with upper case.
 
         When
-        - generating config file
+        - Generating config file.
 
         Then
         - Integration code has arguments as lowercase, but sends the arguments to requests as given.
-        - integration yml, the arguments are lower case.
+        - Integration yml, the arguments are lower case.
         """
         path = tmp_path / 'test-collection.json'
         _testutil_create_postman_collection(dest_path=path, with_request={
@@ -266,6 +266,10 @@ class TestPostmanCodeGen:
             "request": {
                 "method": "GET",
                 "header": [],
+                "body": {
+                    "mode": "raw",
+                    "raw": "{\n    \"A\": 2,\n    \"B\": 3,\n    \"c\": 4\n}"
+                },
                 "url": {
                     "raw": "{{url}}/vtapi/v2/test/{{FOO_A}}?RESOURCE_B=https://www.cobaltstrike.com/",
                     "host": [
@@ -300,12 +304,15 @@ class TestPostmanCodeGen:
         integration_yml = yaml.dump(integration_obj.to_dict())
 
         assert "foo_a = args.get('foo_a')" in integration_code
-        assert "def test_report_request(self, foo_a, resource):" in integration_code
+        assert "def test_report_request(self, foo_a, resource_b, a, b, c)" in integration_code
         assert 'assign_params(RESOURCE_B=resource_b' in integration_code
-        assert "'GET', f'vtapi/v2/test/{foo_a}', params=params, headers=headers)" in integration_code
+        assert "('GET', f'vtapi/v2/test/{foo_a}', params=params, json_data=data, headers=headers)" in integration_code
 
         assert 'name: foo_a' in integration_yml
         assert 'name: resource_b' in integration_yml
+        assert 'name: a\n' in integration_yml
+        assert 'name: b\n' in integration_yml
+        assert 'name: c\n' in integration_yml
 
     def test_apikey_passed_as_header(self, tmpdir):
         """